requirements.txt
fastapi
uvicorn[standard]
pillow
cryptography
psycopg2-binary
qrcode[pil]
pydantic